# swagger.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/change-role": {
      "post": {
        "operationId": "UserController_changeUserRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
...
  },
  "info": {
    "title": "Career Climb",
    "description": "The Career Climb API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          }
        },
        "required": [
          "role"
        ]
      },
...
    }
  }
}